go/types: don't write during sanitizeInfo if nothing has changed
In its final phase, the typechecker walks the types it produces to
ensure that no unexpanded type instances leak through the API. However,
this also walks shared types (such as those in the universe scope),
resulting in a potential data race during concurrent typechecking
passes.

Fix this by simply ensuring that if nothing needs to be sanitized, no
writes occur. Since sanitization is idempotent, this should eliminate
data races.

For #44434

Change-Id: Iadb2e78863efe0e974e69a00e255f26cfaf9386a