spec: add type parameters to the language
[dev.typeparams] spec: add type parameters to the language

This grows the spec (lines of html text counted by wc)
by approx. 15% which seems reasonable given the scope
of changes.

There are a handful of sections (mostly for built-in function)
that still need some prose for type parameter types.

The prose needs wordsmithing and the examples could be better.

Change-Id: Iad3eab4e5696cd5b8562b98c5d2930099c3526f1