cmd/compile: renumber -G flag values
The current -G values were incrementally added as generics support was
plumbed further through the compiler to avoid breaking existing tests
that only worked yet with the earlier stages. Now that we have
reasonably robust end-to-end support for using types2, we can simplify
the flags somewhat to be more useful to end users.

In particular, this CL renumbers the -G values as proposed in #43931:

  -G=0: continue using legacy typechecker
  -G=1: use types2, but without support for generics
  -G=2: use types2, with support for generics

The new -G=2 flag is equivalent to what was formerly -G=3, and
hopefully the new -G=1 flag can be enabled by default for Go 1.17
after we finish reconciling types2's error messages with the existing
regress tests.

While here, also add a -d=typecheckonly debug flag that can be used to
have the compiler gracefully exit after typechecking, so that we can
help distinguish frontend vs backend errors. This also happens to be
needed for smoketest.go, which typechecks at the moment, but does not
yet compile.

Updates #43931.

Change-Id: Id1407bf375bda91d0eb0c28422c5ac33e1760a1a