[dev.fuzz] testing: use exit status 70 for worker errors (not crashes)
If a worker process encounters an error communicating with the
coordinator, or if the setup code reports an error with F.Fail
before calling F.Fuzz, exit with status 70. The coordinator will report
these errors and 'go test' will exit non-zero, but the coordinator
won't record a crasher, since the problem is not in the code being
fuzzed.

Exit codes are chosen somewhat arbitrary, but in the Advanced Bash
Scripting Guide, 70 is "internal software error" which is applicable
here. 70 is also ASCII 'F'.

Change-Id: I1e676e39a7b07c5664efaaa3221d055f55240fff