runtime, time: disable preemption in addtimer
The timerpMask optimization updates a mask of Ps (potentially)
containing timers in pidleget / pidleput. For correctness, it depends on
the assumption that new timers can only be added to a P's own heap.

addtimer violates this assumption if it is preempted after computing pp.
That G may then run on a different P, but adding a timer to the original
P's heap.

Avoid this by disabling preemption while pp is in use.

Other uses of doaddtimer should be OK:

* modtimer: does not add new timers
* cleantimers: does not add new timers
* moveTimers: always moves to the current P
* addAdjustedTimers: does not add new timers
* runtimer: does not add new timers

Fixes #44868

Change-Id: I4a5d080865e854931d0a3a09a51ca36879101d72