crypto/md5: improve ppc64x performance
This is mostly cleanup and simplification.  This removes
many unneeded register moves, loads, and bit twiddlings
which were holdover from porting this from the amd64
version.

This also locates shuffles things around to load an entire
block per computation.  The logical operations now match
the original md5 specification.  This saves a few
instructions.

All in all, this is roughly a 20% improvement.

Change-Id: I269bfa6878966bb4f6a64dc349100f5dc453ab7c