cmd/link: support 32b TLS_LE offsets on PPC64
When using the GCC thread sanitizer, we need larger offsets
as the gcc sanitizer uses quite a bit of thread local space.

Replace the usage of R_PPC64_TPREL16 relocations with a pair
of R_PPC64_TPREL16_HA/LO relocations, and change our internal
R_POWER_TLS_LE relocations to be and use 8 bytes.

Likewise, this illustrated some issues with TLS usage in shared
code.  That is, we should emit an entire TLS IE sequence at once
with one assembly pseudo-op.  This ensures both TLS LE and IE
accesses behave similarly to avoid the issue above.

Notably, this does change the behavior of the objabi.R_POWER_TLS_LE
relocation.  It supports 32b offsets similar to the IE variant.

TODO: is the syntax correct when accessing TLS variables? The
assembler moves the address, not the value into the register.

Change-Id: I1ee9afd0b427cd79888032aa1f60d3e265073e1d