io: fix spelling in documentation for io.Discard
In the process of refactoring ioutil.Discard to io.Discard in
cb0a0f52e67f128c6ad69027c9a8c7a5caf58446 "an" should
have been changed to "a" but was likely missed in the process.

This commit corrects the spelling of the documentation.

This PR will be imported into Gerrit with the title and first
comment (this text) used to generate the subject and body of
the Gerrit change.

Change-Id: I0609c45878291f8f01560efc3f3e6fba191e095b
GitHub-Last-Rev: 315665f2d002e290dabd645d01746d7d92c775e8
GitHub-Pull-Request: golang/go#45190