cmd/compile, internal/abi: add FuncPC intrinsic
When ABI wrappers are used, there are cases where in Go code we
need the PC of the defined function instead of the ABI wrapper.
Currently we work around this by define such functions as
ABIInternal, even if they do not actually follow the internal ABI.

This CL introduces internal/abi.FuncPC as a compiler intrinsic,
which returns the underlying defined function's entry PC if the
argument is a direct reference of a defined function.

As a proof of concept, change runtime.goexit back to ABI0 and use
internal/abi.FuncPC to retrieve its PC.

Updates #44065.

Change-Id: I2d40697f56a925cdeeb0bbd36aac744e25341b2a