[dev.fuzz] internal/fuzz: minimize non-recoverable errors
Assuming that this works for non-recoverable errors, there
will likely be a follow-up CL which refactors the minimization
for recoverable errors to use the same RPC flow (since that
more easily allows the worker to tell the coordinator that
it's minimizing and shouldn't send more inputs to other workers
to fuzz).

There is still a bug where every panic is logged to the command
line, even while minimizing it. That can get pretty noisy.
There is a TODO to address this in a follow-up change.

Change-Id: I32ac7cec4abe2d4c345c0ee77315233047efb1fb