database/sql,net,net/http,os/exec: Replace ctx.Done() with ctx.Err()
Prior to this commit, we performed non-blocking tests for cancellation
by receiving from ctx.Done() in a select block with a default label
like:

  select {
    case <-ctx.Done():
      // handle cancellation
    default:
  }

Unless we're multiplexing over multiple channels, a better way to test
if a context is canceled is to test the result of ctx.Err()

This commit replaces all of the select{} based non-blocking cancellation
checks with a call to ctx.Err().

The src/context BenchmarkCheckCanceled benchmark demonstrates that
calling ctx.Err() is faster and, in general, it is probably clearer to
simply check ctx.Err() than using a select block.

Change-Id: I1de2fdca200afd81e3bd1c9ade8ea02e19324ae5
GitHub-Last-Rev: d6f2b3e38ccb48e3649ece72556522b1e26715e4
GitHub-Pull-Request: golang/go#43189