go/parser: add a SkipResolution method to bypass object resolution
Parser object resolution is an auxiliary feature in which the parser
attempts to resolve identifiers to their declarations. In functionality,
it significantly overlaps with go/types and in fact cannot be correctly
computed at parse-time without type information (specifically: the
declaration of k in the composite lit c{k: v} is undecidable). Due to
these limitations, it is of limited utility and is often unused by tools
that only need syntax information, or which rely on go/types for object
resolution.

Now that object resolution is isolated as a post-processing pass, it is
trivial to offer a parser mode that skips it entirely. This CL adds that
mode.

Fixes golang/go#45104

Change-Id: I5a2c05437e298964ad2039e1ff98e63d6efbd1af